fix + feat: update GTB spline implementation + tests
A rewrite of GTB-spline implementation where, if all canonical spaces
are the same, then the corresponding spline space is initialized via the
B-spline structure and not the unstructured spline structure. Also, the
local section spaces for a B-spline space now are not limited to
Bernstein but can also be Lagrange. In this way, the B-spline space is
not really a B-spline space, but it is rather a univariate finite
element patch object.